Move MutationTypesEnum
